feat: modulation
